requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.32.0
django-filter==1.0.4
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
PyJWT==1.5.2
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
